Rename isNode.ts to guards.ts
