+++ packages/arbor-store/src/Arbor.ts
@@ -6,7 +6,7 @@ import NodeMapHandler from "./NodeMapHandler"
 import Path from "./Path"
 import Subscribers, { Subscriber, Unsubscribe } from "./Subscribers"
 import { DetachedNodeError, NotAnArborNodeError } from "./errors"
-import isNode from "./isNode"
+import isNode from "./guards"
 import { ArborProxiable } from "./isProxiable"
 import mutate, { Mutation, MutationMetadata } from "./mutate"
 import { notifyAffectedSubscribers } from "./notifyAffectedSubscribers"
