+++ packages/arbor-store/src/guards.ts
