+++ packages/arbor-store/src/index.ts
@@ -16,7 +16,7 @@ export type { Record } from "./Repository"
 export { default as Subscribers } from "./Subscribers"
 export type { MutationEvent, Subscriber, Unsubscribe } from "./Subscribers"
 export * from "./errors"
-export { isArborNode, default as isNode } from "./isNode"
+export { isArborNode, default as isNode } from "./guards"
 export { ArborProxiable, default as isProxiable } from "./isProxiable"
 export type { Mutation } from "./mutate"
 export * from "./utilities"
