+++ packages/arbor-store/src/utilities.ts
@@ -1,7 +1,7 @@
 import { ArborNode } from "./Arbor"
 import Path from "./Path"
 import { ArborError, DetachedNodeError, NotAnArborNodeError } from "./errors"
-import isNode from "./isNode"
+import isNode from "./guards"
 
 /**
  * Detaches a given ArborNode from the state tree.
